Cloud-Automation/GCP/LAB01-Compute-Instance-Creation/requirements.txt
google-cloud-compute==1.14.1
google-api-python-client==2.108.0
google-auth==2.23.4
python-dotenv==1.0.0 